+++ src/Git_Remote_Updater/GRU_Freemius.php
@@ -10,6 +10,8 @@
 
 namespace Fragen\Git_Remote_Updater;
 
+use Fragen\Git_Updater\Shim;
+
 /*
  * Exit if called directly.
  */
@@ -40,6 +42,10 @@ class GRU_Freemius {
 				global $gru_fs;
 
 				if ( ! isset( $gru_fs ) ) {
+
+					// Init Freemius SDK.
+					require_once Shim::dirname( __DIR__, 2 ) . '/vendor/freemius/wordpress-sdk/start.php';
+
 					$gru_fs = fs_dynamic_init(
 						[
 							'id'               => '8312',
@@ -83,6 +89,6 @@ class GRU_Freemius {
 	 * @return string
 	 */
 	public function add_icon() {
-		return dirname( __DIR__, 2 ) . '/assets/icon.svg';
+		return Shim::dirname( __DIR__, 2 ) . '/assets/icon.svg';
 	}
 }